Extracting generic helpers into a utils package
CleanString, FormatKey, NormalizeNumericString, NormalizeTaxPercent,
FormatDate, ParseDate, RescaleToCurrency, and CalculateRequiredPrecision
have no UBL-specific logic -- they're generic string/number/date helpers
that happened to live alongside the UBL wire model. Moving them to their
own package (exported, since a separate package can't reach unexported
identifiers) makes the split between "generic helper" and "UBL
conversion logic" explicit, and gives country overlays a stable, focused
import for just the helpers, without pulling in the whole UBL model.

Left in place: goblUnitFromUNECE, ParseNote/FormatNote, and
buildTaxCategoryKey (UBL/UNTDID-specific conventions), InvoiceNumber
(UBL ID-field convention), and contactName (private, tightly coupled to
NewParty's assembly flow).
