+++ utils/dates_test.go
@@ -1,0 +1,34 @@
+package utils
+
+import (
+	"testing"
+
+	"github.com/stretchr/testify/assert"
+)
+
+// Define tests for the ParseDate function
+func TestParseDate(t *testing.T) {
+	tests := []struct {
+		name        string
+		input       string
+		expected    string
+		expectError bool
+	}{
+		{"Valid date", "2023-05-15", "2023-05-15", false},
+		{"Invalid date", "2023-13-45", "", true},
+		{"Empty string", "", "", true},
+	}
+
+	for _, tt := range tests {
+		t.Run(tt.name, func(t *testing.T) {
+			result, err := ParseDate(tt.input)
+			if tt.expectError {
+				assert.Error(t, err)
+				assert.Empty(t, result)
+			} else {
+				assert.NoError(t, err)
+				assert.Equal(t, tt.expected, result.String())
+			}
+		})
+	}
+}
